% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structures.R
\name{as.list.tracks}
\alias{as.list.tracks}
\title{Convert from Tracks to List}
\usage{
\method{as.list}{tracks}(x, ...)
}
\arguments{
\item{x}{the \code{tracks} object to be coerced to a list.}

\item{...}{further arguments to be passed from or to other methods.}
}
\value{
A generic list of single tracks, where each track is a matrix with 
\code{t/delta.t} rows and 4 columns. This looks a lot like a tracks object,
except that its class is not "tracks" anymore.
}
\description{
Coerces a \code{tracks} object to a list.
}
