% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltrackR-package.R
\name{cheatsheet}
\alias{cheatsheet}
\title{Open the package cheat sheet}
\usage{
cheatsheet(opencmd = c("open", "xdg-open", "gvfs-open", "gio open"))
}
\arguments{
\item{opencmd}{The command used to open pdfs from the command line.}
}
\description{
Running this function will open the package cheat sheet (a pdf) via a call to
\code{system()}.
}
