% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{maxTrackLength}
\alias{maxTrackLength}
\title{Length of Longest Track}
\usage{
maxTrackLength(x)
}
\arguments{
\item{x}{the \code{tracks} object the tracks in which are to be considered.}
}
\value{
The maximum number of rows of a track in \code{x}
}
\description{
Determines the maximum number of positions over the tracks in \code{x}.
}

