% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalizeToDuration}
\alias{normalizeToDuration}
\title{Normalize a Measure to Track Duration}
\usage{
normalizeToDuration(x)
}
\arguments{
\item{x}{a track measure (see \link{TrackMeasures}).}
}
\value{
A function that computes the input measure for a given track
and returns the result divided by the track's duration.
}
\description{
Returns a measure that divides the input measure by the duration of its 
input track.
}
\examples{
## normalizeToDuration(displacement) can be used as an indicator 
## for the motion's efficiency
sapply(TCells, normalizeToDuration(displacement))
}

