% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sort.tracks}
\alias{sort.tracks}
\title{Sort Track Positions by Time}
\usage{
\method{sort}{tracks}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{the \emph{tracks} object whose tracks are to be sorted by time.}

\item{decreasing}{logical.  Should the sort be increasing or decreasing? 
Provided only for consistency with the generic sort method. The positions in
each track should be sorted in increasing time order.}

\item{...}{further arguments to be passed on to \code{order}.}
}
\value{
A \emph{tracks object} that contains the tracks from the input object 
sorted by time is returned.
}
\description{
Sorts the positions in each track in a \emph{tracks} object by time.
}
\details{
Sorts the positions of each track (represented as a data frame) in the 
\emph{tracks} object by time (given in the column t).
}

