% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{splitTrack}
\alias{splitTrack}
\title{Split Track into Multiple Tracks}
\usage{
splitTrack(x, positions, id = NULL, min.length = 2)
}
\arguments{
\item{x}{the input track (a data frame or a matrix).}

\item{positions}{a vector of positive integers, given in ascending order.}

\item{id}{a string used to identify the resulting tracks; for instance,
if \code{id="X"}, then the resulting tracks are named X_1, X_2 and so forth.
Otherwise, they are simply labelled with integer numbers.}

\item{min.length}{nonnegative integer. Resulting tracks that have fewer positions than
the value of this parameter are dropped.}
}
\description{
Split Track into Multiple Tracks
}

