% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtracks.R
\name{subtracksByTime}
\alias{subtracksByTime}
\title{Extract Subtracks Starting at a Specific Time}
\usage{
subtracksByTime(X, t, i, digits = 5)
}
\arguments{
\item{X}{Tracks object to obtain subtracks from}

\item{t}{Timepoint at which the subtracks should start}

\item{i}{Subtrack length (in number of steps)}

\item{digits}{For numeric reasons, timepoints are rounded to this number of digits
(see details).}
}
\value{
A \emph{tracks} object is returned which contains all the subtracks
of any track in the input \emph{tracks} object that consist of exactly `i`
segments and start at the given timepoint t.
}
\description{
Obtain all subtracks of i steps (i+1 positions) starting at a given timepoint t.
}
\details{
Timepoints are rounded to a given number of digits for numerical reasons;
otherwise timepoints that are equal can seem like a different number. The given t
is then retrieved for all tracks in X that contain that timepoint, and any subtracks
starting from that time that have i steps are returned.
}
\examples{
## Get and plot all the single steps (i=1) starting at the third timepoint in the T cell tracks.
subT <- subtracksByTime( TCells, timePoints(TCells)[3], 1 )
plot( subT )

}
\seealso{
\code{\link{subtracks}} to extract all subtracks of a given length,
\code{\link{prefixes}} to extract all subtracks of a given length starting
from the first coordinate in each track, \code{\link{selectSteps}} to extract
single steps starting at a fixed timepoint from a subset of trackids, and
\code{\link{timePoints}} to return all timepoints occurring in the dataset.
}
