% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{timePoints}
\alias{timePoints}
\title{Find All Unique Time Points in a Track Dataset}
\usage{
timePoints(X)
}
\arguments{
\item{X}{a tracks object.}
}
\value{
A numeric vector of unique timepoints.
}
\description{
Return a vector of all the timepoints t found in any of the matrices in the
tracks object.
}
\examples{
## Get all timepoints in the T cell dataset
tp <- timePoints( TCells )

}
