% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle-functions.R
\name{vecAngle}
\alias{vecAngle}
\title{Angle Between Two Vectors}
\usage{
vecAngle(a, b, degrees = TRUE)
}
\description{
a and b can be numeric vectors or matrices in which each row represents a numeric vector.
In the last case, one angle is returned for each row.
}
