% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wrapTrack}
\alias{wrapTrack}
\title{Create Track Object from Single Track}
\usage{
wrapTrack(x)
}
\arguments{
\item{x}{the input track.}
}
\value{
A list of class \code{tracks} containing only the input track \code{x}, which
is assigned the name "1".
}
\description{
Makes a \code{tracks} object containing the given track.
}

