% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{as.list.tracks}
\alias{as.list.tracks}
\title{Convert from Tracks to List}
\usage{
\method{as.list}{tracks}(x, ...)
}
\arguments{
\item{x}{the \code{tracks} object to be coerced to a list.}

\item{...}{further arguments to be passed from or to other methods.}
}
\description{
Coerces a \code{tracks} object to a list.
}

