% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-set-transformations.R
\name{normalizeTracks}
\alias{normalizeTracks}
\title{Normalize Tracks}
\usage{
normalizeTracks(x)
}
\arguments{
\item{x}{the input \code{tracks} object.}
}
\description{
Translates each track in a given set of tracks such that the 
first position is the origin.
}
\examples{
## normalization of Neutrophil data reveals upward motion
plot( normalizeTracks( Neutrophils ) )
}

