% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prefixes}
\alias{prefixes}
\title{Get Track Prefixes}
\usage{
prefixes(x, i)
}
\arguments{
\item{x}{a single track or a \code{tracks} object.}

\item{i}{subtrack length. A single integer, lists are not supported.}
}
\description{
Creates a \code{tracks} object consisting of all prefixes (i.e., subtracks
starting with the first position of a track) of `x`
with `i` segments (i.e., `i`+1 positions).
}
\details{
This function behaves exactly like \code{\link{subtracks}} except
that only subtracks starting from the first position are considered.
}

